# api/openapi.yaml
openapi: 3.1.0
info:
  title: Sheily local sovereign agent
  description: Agente cognitivo local con autoridades independientes y ejecución Rust.
  version: 0.3.0
paths:
...
  /v1/auth/login:
    post:
      summary: Login
      operationId: login_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginRequest:
      properties:
        username:
          type: string
          maxLength: 128
          minLength: 1
          title: Username
        password:
          type: string
          maxLength: 512
          minLength: 1
          title: Password
      additionalProperties: false
      type: object
      required:
      - username
      - password
      title: LoginRequest
...
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          const: bearer
          title: Token Type
          default: bearer
        expires_in:
          type: integer
          title: Expires In
        user_id:
          type: string
          title: User Id
        role:
          type: string
          title: Role
      additionalProperties: false
      type: object
      required:
      - access_token
      - expires_in
      - user_id
      - role
      title: TokenResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError